docs(switch): add playground example
