+++ skills/carbon-react/components/switch.md
@@ -64,230 +64,4 @@ description: Carbon Switch component props and usage examples.
 | warning | string \| boolean \| undefined | No |  | Yes | This prop is no longer supported. |  |  |
 
 ## Examples
-### Default
-
-**Render**
-
-```tsx
-() => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Toggle notifications"
-      inputHint="Hint text"
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-}
-```
-
-
-### Checked
-
-**Render**
-
-```tsx
-() => {
-  const [checked, setChecked] = useState(true);
-  return (
-    <Switch
-      label="Toggle notifications"
-      inputHint="Hint text"
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-}
-```
-
-
-### Disabled
-
-**Render**
-
-```tsx
-() => (
-  <Switch
-    label="Toggle notifications"
-    inputHint="Hint text"
-    checked={false}
-    disabled
-    onChange={() => {}}
-  />
-)
-```
-
-
-### Disabled (checked)
-
-**Render**
-
-```tsx
-() => (
-  <Switch
-    label="Toggle notifications"
-    checked
-    disabled
-    inputHint="Hint text"
-    onChange={() => {}}
-  />
-)
-```
-
-
-### Required
-
-**Render**
-
-```tsx
-() => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Toggle notifications"
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-      required
-    />
-  );
-}
-```
-
-
-### Large size
-
-**Render**
-
-```tsx
-() => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Toggle notifications"
-      inputHint="Hint text"
-      size="large"
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-}
-```
-
-
-### Label inline
-
-**Render**
-
-```tsx
-() => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Toggle notifications"
-      labelInline
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-}
-```
-
-
-### Label inline with hint
-
-**Render**
-
-```tsx
-() => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Label"
-      inputHint="Hint text"
-      labelInline
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-}
-```
-
-
-### Loading
-
-**Render**
-
-```tsx
-() => (
-  <>
-    <Switch
-      label="Toggle notifications"
-      checked={false}
-      loading
-      onChange={() => {}}
-    />
-
-    <Switch
-      label="Toggle notifications"
-      checked
-      loading
-      onChange={() => {}}
-      ml="8px"
-    />
-
-    <Switch
-      size="large"
-      label="Toggle notifications"
-      checked={false}
-      loading
-      onChange={() => {}}
-      ml="8px"
-    />
-
-    <Switch
-      size="large"
-      label="Toggle notifications"
-      checked
-      loading
-      onChange={() => {}}
-      ml="8px"
-    />
-  </>
-)
-```
-
-
-### Loading with custom processingLabel
-
-**Render**
-
-```tsx
-() => (
-  <Switch
-    label="Toggle notifications"
-    checked={false}
-    loading
-    processingLabel="Saving changes..."
-    onChange={() => {}}
-  />
-)
-```
-
-
-### Loading with processingLabel below switch
-
-**Render**
-
-```tsx
-() => (
-  <Switch
-    label="Toggle notifications"
-    checked={false}
-    loading
-    processingLabelBelowSwitch
-    onChange={() => {}}
-  />
-)
-```
-
+No Storybook examples found.
