+++ src/components/switch/switch.stories.tsx
@@ -12,6 +12,25 @@ const meta: Meta<typeof Switch> = {
   component: Switch,
   argTypes: {
     ...styledSystemProps,
+    size: {
+      options: ["small", "large"],
+      control: { type: "radio" },
+    },
+    label: {
+      control: "text",
+    },
+    inputHint: {
+      control: "text",
+    },
+    disabled: {
+      control: "boolean",
+    },
+    loading: {
+      control: "boolean",
+    },
+    required: {
+      control: "boolean",
+    },
   },
   parameters: {
     themeProvider: { chromatic: { theme: "sage" } },
@@ -22,164 +41,29 @@ const meta: Meta<typeof Switch> = {
 export default meta;
 type Story = StoryObj<typeof Switch>;
 
-export const Default: Story = () => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Toggle notifications"
-      inputHint="Hint text"
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
+export const Playground: Story = {
+  render: (args) => {
+    const [checked, setChecked] = useState(false);
+    return (
+      <Switch
+        {...args}
+        checked={checked}
+        onChange={(e) => setChecked(e.target.checked)}
+      />
+    );
+  },
+  args: {
+    label: "Toggle notifications",
+    inputHint: "Hint text",
+    disabled: false,
+    loading: false,
+    required: false,
+    size: "small",
+    processingLabel: "Processing...",
+    processingLabelBelowSwitch: false,
+    labelInline: false,
+    labelSpacing: 1,
+    labelWidth: 30,
+  },
 };
-Default.storyName = "Default";
-
-export const Checked: Story = () => {
-  const [checked, setChecked] = useState(true);
-  return (
-    <Switch
-      label="Toggle notifications"
-      inputHint="Hint text"
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-};
-Checked.storyName = "Checked";
-
-export const Disabled: Story = () => (
-  <Switch
-    label="Toggle notifications"
-    inputHint="Hint text"
-    checked={false}
-    disabled
-    onChange={() => {}}
-  />
-);
-Disabled.storyName = "Disabled";
-
-export const DisabledChecked: Story = () => (
-  <Switch
-    label="Toggle notifications"
-    checked
-    disabled
-    inputHint="Hint text"
-    onChange={() => {}}
-  />
-);
-DisabledChecked.storyName = "Disabled (checked)";
-
-export const Required: Story = () => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Toggle notifications"
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-      required
-    />
-  );
-};
-Required.storyName = "Required";
-
-export const LargeSize: Story = () => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Toggle notifications"
-      inputHint="Hint text"
-      size="large"
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-};
-LargeSize.storyName = "Large size";
-
-export const LabelInline: Story = () => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Toggle notifications"
-      labelInline
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-};
-LabelInline.storyName = "Label inline";
-
-export const LabelInlineWithHint: Story = () => {
-  const [checked, setChecked] = useState(false);
-  return (
-    <Switch
-      label="Label"
-      inputHint="Hint text"
-      labelInline
-      checked={checked}
-      onChange={(e) => setChecked(e.target.checked)}
-    />
-  );
-};
-LabelInlineWithHint.storyName = "Label inline with hint";
-
-export const Loading: Story = () => (
-  <>
-    <Switch
-      label="Toggle notifications"
-      checked={false}
-      loading
-      onChange={() => {}}
-    />
-
-    <Switch
-      label="Toggle notifications"
-      checked
-      loading
-      onChange={() => {}}
-      ml="8px"
-    />
-
-    <Switch
-      size="large"
-      label="Toggle notifications"
-      checked={false}
-      loading
-      onChange={() => {}}
-      ml="8px"
-    />
-
-    <Switch
-      size="large"
-      label="Toggle notifications"
-      checked
-      loading
-      onChange={() => {}}
-      ml="8px"
-    />
-  </>
-);
-Loading.storyName = "Loading";
-
-export const LoadingCustomLabel: Story = () => (
-  <Switch
-    label="Toggle notifications"
-    checked={false}
-    loading
-    processingLabel="Saving changes..."
-    onChange={() => {}}
-  />
-);
-LoadingCustomLabel.storyName = "Loading with custom processingLabel";
-
-export const LoadingLabelBelow: Story = () => (
-  <Switch
-    label="Toggle notifications"
-    checked={false}
-    loading
-    processingLabelBelowSwitch
-    onChange={() => {}}
-  />
-);
-LoadingLabelBelow.storyName = "Loading with processingLabel below switch";
+Playground.storyName = "Playground";
